+++ packages/google_maps_flutter/google_maps_flutter/example/lib/tile_overlay.dart
@@ -13,7 +13,8 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'page.dart';
 
 class TileOverlayPage extends GoogleMapExampleAppPage {
-  const TileOverlayPage() : super(const Icon(Icons.map), 'Tile overlay');
+  const TileOverlayPage({Key? key})
+      : super(const Icon(Icons.map), 'Tile overlay', key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -22,7 +23,7 @@ class TileOverlayPage extends GoogleMapExampleAppPage {
 }
 
 class TileOverlayBody extends StatefulWidget {
-  const TileOverlayBody();
+  const TileOverlayBody({Key? key}) : super(key: key);
 
   @override
   State<StatefulWidget> createState() => TileOverlayBodyState();
@@ -90,16 +91,16 @@ class TileOverlayBodyState extends State<TileOverlayBody> {
           ),
         ),
         TextButton(
-          child: const Text('Add tile overlay'),
           onPressed: _addTileOverlay,
+          child: const Text('Add tile overlay'),
         ),
         TextButton(
-          child: const Text('Remove tile overlay'),
           onPressed: _removeTileOverlay,
+          child: const Text('Remove tile overlay'),
         ),
         TextButton(
-          child: const Text('Clear tile cache'),
           onPressed: _clearTileCache,
+          child: const Text('Clear tile cache'),
         ),
       ],
     );