+++ packages/in_app_purchase/in_app_purchase_android/example/lib/main.dart
@@ -37,7 +37,7 @@ const List<String> _kProductIds = <String>[
 
 class _MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() => _MyAppState();
+  State<_MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<_MyApp> {
@@ -264,7 +264,6 @@ class _MyAppState extends State<_MyApp> {
                     },
                     icon: const Icon(Icons.upgrade))
                 : TextButton(
-                    child: Text(productDetails.price),
                     style: TextButton.styleFrom(
                       backgroundColor: Colors.green[800],
                       primary: Colors.white,
@@ -297,6 +296,7 @@ class _MyAppState extends State<_MyApp> {
                             purchaseParam: purchaseParam);
                       }
                     },
+                    child: Text(productDetails.price),
                   ));
       },
     ));
@@ -337,9 +337,9 @@ class _MyAppState extends State<_MyApp> {
       const Divider(),
       GridView.count(
         crossAxisCount: 5,
-        children: tokens,
         shrinkWrap: true,
         padding: const EdgeInsets.all(16.0),
+        children: tokens,
       )
     ]));
   }