+++ packages/local_auth/local_auth/lib/src/local_auth.dart
@@ -11,12 +11,9 @@
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-import 'package:local_auth/src/types/error_codes.dart';
 import 'package:local_auth_android/local_auth_android.dart';
 import 'package:local_auth_ios/local_auth_ios.dart';
 import 'package:local_auth_platform_interface/local_auth_platform_interface.dart';
-import 'package:local_auth_platform_interface/types/auth_messages.dart';
-import 'package:local_auth_platform_interface/types/biometric_type.dart';
 
 /// A Flutter plugin for authenticating the user identity locally.
 class LocalAuthentication {