+++ packages/webview_flutter/webview_flutter/example/integration_test/webview_flutter_test.dart
@@ -14,10 +14,8 @@ import 'dart:typed_data';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:webview_flutter/platform_interface.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 Future<void> main() async {
@@ -1308,7 +1306,8 @@ Future<String> _runJavascriptReturningResult(
 
 class ResizableWebView extends StatefulWidget {
   const ResizableWebView(
-      {required this.onResize, required this.onPageFinished});
+      {Key? key, required this.onResize, required this.onPageFinished})
+      : super(key: key);
 
   final JavascriptMessageHandler onResize;
   final VoidCallback onPageFinished;