+++ script/tool/test/util.dart
@@ -13,12 +13,15 @@ import 'package:flutter_plugin_tools/src/common/core.dart';
 import 'package:flutter_plugin_tools/src/common/file_utils.dart';
 import 'package:flutter_plugin_tools/src/common/plugin_utils.dart';
 import 'package:flutter_plugin_tools/src/common/process_runner.dart';
+import 'package:flutter_plugin_tools/src/common/repository_package.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:platform/platform.dart';
 import 'package:quiver/collection.dart';
 
 import 'mocks.dart';
+
+export 'package:flutter_plugin_tools/src/common/repository_package.dart';
 
 /// Returns the exe name that command will use when running Flutter on
 /// [platform].
@@ -65,6 +68,20 @@ class PlatformDetails {
   final bool hasDartCode;
 }
 
+/// Returns the 'example' directory for [package].
+///
+/// This is deliberately not a method on [RepositoryPackage] since actual tool
+/// code should essentially never need this, and instead be using
+/// [RepositoryPackage.getExamples] to avoid assuming there's a single example
+/// directory. However, needing to construct paths with the example directory
+/// is very common in test code.
+///
+/// This returns a Directory rather than a RepositoryPackage because there is no
+/// guarantee that the returned directory is a package.
+Directory getExampleDir(RepositoryPackage package) {
+  return package.directory.childDirectory('example');
+}
+
 /// Creates a plugin package with the given [name] in [packagesDirectory].
 ///
 /// [platformSupport] is a map of platform string to the support details for
@@ -72,8 +89,7 @@ class PlatformDetails {
 ///
 /// [extraFiles] is an optional list of plugin-relative paths, using Posix
 /// separators, of extra files to create in the plugin.
-// TODO(stuartmorgan): Convert the return to a RepositoryPackage.
-Directory createFakePlugin(
+RepositoryPackage createFakePlugin(
   String name,
   Directory parentDirectory, {
   List<String> examples = const <String>['example'],
@@ -83,7 +99,7 @@ Directory createFakePlugin(
   String? version = '0.0.1',
   String flutterConstraint = '>=2.5.0',
 }) {
-  final Directory pluginDirectory = createFakePackage(name, parentDirectory,
+  final RepositoryPackage package = createFakePackage(name, parentDirectory,
       isFlutter: true,
       examples: examples,
       extraFiles: extraFiles,
@@ -91,7 +107,7 @@ Directory createFakePlugin(
       flutterConstraint: flutterConstraint);
 
   createFakePubspec(
-    pluginDirectory,
+    package,
     name: name,
     isFlutter: true,
     isPlugin: true,
@@ -100,15 +116,20 @@ Directory createFakePlugin(
     flutterConstraint: flutterConstraint,
   );
 
-  return pluginDirectory;
+  return package;
 }
 
 /// Creates a plugin package with the given [name] in [packagesDirectory].
 ///
 /// [extraFiles] is an optional list of package-relative paths, using unix-style
 /// separators, of extra files to create in the package.
-// TODO(stuartmorgan): Convert the return to a RepositoryPackage.
-Directory createFakePackage(
+///
+/// If [includeCommonFiles] is true, common but non-critical files like
+/// CHANGELOG.md and AUTHORS will be included.
+///
+/// If non-null, [directoryName] will be used for the directory instead of
+/// [name].
+RepositoryPackage createFakePackage(
   String name,
   Directory parentDirectory, {
   List<String> examples = const <String>['example'],
@@ -116,37 +137,42 @@ Directory createFakePackage(
   bool isFlutter = false,
   String? version = '0.0.1',
   String flutterConstraint = '>=2.5.0',
+  bool includeCommonFiles = true,
+  String? directoryName,
+  String? publishTo,
 }) {
-  final Directory packageDirectory = parentDirectory.childDirectory(name);
-  packageDirectory.createSync(recursive: true);
+  final RepositoryPackage package =
+      RepositoryPackage(parentDirectory.childDirectory(directoryName ?? name));
+  package.directory.createSync(recursive: true);
 
-  createFakePubspec(packageDirectory,
+  package.libDirectory.createSync();
+  createFakePubspec(package,
       name: name,
       isFlutter: isFlutter,
       version: version,
       flutterConstraint: flutterConstraint);
-  createFakeCHANGELOG(packageDirectory, '''
+  if (includeCommonFiles) {
+    createFakeCHANGELOG(package, '''
 ## $version
   * Some changes.
   ''');
-  createFakeAuthors(packageDirectory);
+    createFakeAuthors(package);
+  }
 
   if (examples.length == 1) {
-    final Directory exampleDir = packageDirectory.childDirectory(examples.first)
-      ..createSync();
-    createFakePubspec(exampleDir,
-        name: '${name}_example',
+    createFakePackage('${name}_example', package.directory,
+        directoryName: examples.first,
+        examples: <String>[],
+        includeCommonFiles: false,
         isFlutter: isFlutter,
         publishTo: 'none',
         flutterConstraint: flutterConstraint);
   } else if (examples.isNotEmpty) {
-    final Directory exampleDir = packageDirectory.childDirectory('example')
-      ..createSync();
-    for (final String example in examples) {
-      final Directory currentExample = exampleDir.childDirectory(example)
-        ..createSync();
-      createFakePubspec(currentExample,
-          name: example,
+    final Directory examplesDirectory = getExampleDir(package)..createSync();
+    for (final String exampleName in examples) {
+      createFakePackage(exampleName, examplesDirectory,
+          examples: <String>[],
+          includeCommonFiles: false,
           isFlutter: isFlutter,
           publishTo: 'none',
           flutterConstraint: flutterConstraint);
@@ -155,16 +181,16 @@ Directory createFakePackage(
 
   final p.Context posixContext = p.posix;
   for (final String file in extraFiles) {
-    childFileWithSubcomponents(packageDirectory, posixContext.split(file))
+    childFileWithSubcomponents(package.directory, posixContext.split(file))
         .createSync(recursive: true);
   }
 
-  return packageDirectory;
+  return package;
 }
 
-void createFakeCHANGELOG(Directory parent, String texts) {
-  parent.childFile('CHANGELOG.md').createSync();
-  parent.childFile('CHANGELOG.md').writeAsStringSync(texts);
+void createFakeCHANGELOG(RepositoryPackage package, String texts) {
+  package.changelogFile.createSync();
+  package.changelogFile.writeAsStringSync(texts);
 }
 
 /// Creates a `pubspec.yaml` file with a flutter dependency.
@@ -173,13 +199,13 @@ void createFakeCHANGELOG(Directory parent, String texts) {
 /// that platform. If empty, no `plugin` entry will be created unless `isPlugin`
 /// is set to `true`.
 void createFakePubspec(
-  Directory parent, {
+  RepositoryPackage package, {
   String name = 'fake_package',
   bool isFlutter = true,
   bool isPlugin = false,
   Map<String, PlatformDetails> platformSupport =
       const <String, PlatformDetails>{},
-  String publishTo = 'http://no_pub_server.com',
+  String? publishTo,
   String? version,
   String dartConstraint = '>=2.0.0 <3.0.0',
   String flutterConstraint = '>=2.5.0',
@@ -219,9 +245,16 @@ flutter:
     }
   }
 
-  String yaml = '''
+  // Default to a fake server to avoid ever accidentally publishing something
+  // from a test. Does not use 'none' since that changes the behavior of some
+  // commands.
+  final String publishToSection =
+      'publish_to: ${publishTo ?? 'http://no_pub_server.com'}';
+
+  final String yaml = '''
 name: $name
 ${(version != null) ? 'version: $version' : ''}
+$publishToSection
 
 $environmentSection
 
@@ -230,19 +263,13 @@ $dependenciesSection
 $pluginSection
 ''';
 
-  if (publishTo.isNotEmpty) {
-    yaml += '''
-publish_to: $publishTo # Hardcoded safeguard to prevent this from somehow being published by a broken test.
-''';
-  }
-  parent.childFile('pubspec.yaml').createSync();
-  parent.childFile('pubspec.yaml').writeAsStringSync(yaml);
+  package.pubspecFile.createSync();
+  package.pubspecFile.writeAsStringSync(yaml);
 }
 
-void createFakeAuthors(Directory parent) {
-  final File authorsFile = parent.childFile('AUTHORS');
-  authorsFile.createSync();
-  authorsFile.writeAsStringSync('Google Inc.');
+void createFakeAuthors(RepositoryPackage package) {
+  package.authorsFile.createSync();
+  package.authorsFile.writeAsStringSync('Google Inc.');
 }
 
 String _pluginPlatformSection(
@@ -292,14 +319,14 @@ String _pluginPlatformSection(
   return entry;
 }
 
-typedef _ErrorHandler = void Function(Error error);
+typedef ErrorHandler = void Function(Error error);
 
 /// Run the command [runner] with the given [args] and return
 /// what was printed.
 /// A custom [errorHandler] can be used to handle the runner error as desired without throwing.
 Future<List<String>> runCapturingPrint(
     CommandRunner<void> runner, List<String> args,
-    {_ErrorHandler? errorHandler}) async {
+    {ErrorHandler? errorHandler}) async {
   final List<String> prints = <String>[];
   final ZoneSpecification spec = ZoneSpecification(
     print: (_, __, ___, String message) {