+++ script/tool/CHANGELOG.md
@@ -322,7 +322,7 @@ and `firebase-test-lab`.
 
 ## v.0.0.36+2
 
-- Default to showing podspec lint warnings
+- Default to showing podspec lint warnings.
 
 ## v.0.0.36+1